Удалил лишние сервисы и broadcastReceiver-ы.
