+++ app/src/main/java/com/example/weather/view/WeatherActivity.kt
@@ -2,11 +2,14 @@ package com.example.weather.view
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
+import android.content.IntentFilter
+import android.net.ConnectivityManager
 import android.os.Build
 import android.os.Bundle
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
+import com.example.weather.broadcastReceiver.ConnectivityBroadcastReceiver
 import com.example.weather.databinding.WeatherActivityBinding
 import com.example.weather.utils.CHANNEL_ID
 import com.example.weather.utils.CHANNEL_NAME
@@ -17,6 +20,9 @@ class WeatherActivity : AppCompatActivity() {
 
     private var _binding:WeatherActivityBinding? = null
     private val binding get() = _binding!!
+
+    // создание broadcast-ресивера для приема и обработки данных о изменении подключения к сети
+    private val connectivityBroadcastReceiver = ConnectivityBroadcastReceiver()
 
     @RequiresApi(Build.VERSION_CODES.N)
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -39,6 +45,20 @@ class WeatherActivity : AppCompatActivity() {
                 .replace(binding.container.id, CityListFragment.newInstance())
                 .commitNow()
         }
+    }
+
+    override fun onStart() {
+        // регистрируем connectivityBroadcastReceiver
+            @Suppress("DEPRECATION")
+            registerReceiver(connectivityBroadcastReceiver, IntentFilter(
+                ConnectivityManager.CONNECTIVITY_ACTION))
+        super.onStart()
+    }
+
+    override fun onStop() {
+        unregisterReceiver(connectivityBroadcastReceiver)
+        ConnectivityBroadcastReceiver.isFirstReceive=true
+        super.onStop()
     }
 
     private fun initNotificationChannel() {
