--- app/src/main/res/layout/fragment_thread.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:background="@color/teal_200">
-    <LinearLayout
-        android:id="@+id/mainContainer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:padding="@dimen/margin_15_dp">
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="calculate"
-                android:textSize="@dimen/feels_like_text_size" />
-            <androidx.appcompat.widget.AppCompatEditText
-                android:id="@+id/editText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:ems="2"
-            android:gravity="center_horizontal"
-            android:inputType="number"
-            android:text="1"
-            android:textSize="@dimen/feels_like_text_size" />
-            <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:gravity="center_horizontal"
-                tools:text="0"
-                android:textSize="@dimen/feels_like_text_size" />
-        </LinearLayout>
-        <Button
-            android:id="@+id/calcThreadBtn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:text="calc_in_thread"
-            android:textSize="@dimen/feels_like_text_size" />
-        <Button
-            android:id="@+id/calcThreadHandler"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:text="handler_thread"
-            android:textSize="@dimen/feels_like_text_size" />
-        <Button
-            android:id="@+id/serviceButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:text="service"
-        android:textSize="@dimen/feels_like_text_size" />
-        <Button
-            android:id="@+id/serviceWithBroadcastButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:text="service_broadcast"
-            android:textSize="@dimen/feels_like_text_size" />
-
-    </LinearLayout>
-</ScrollView>
